виндовс_plan_win_bot/app/requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosqlite==0.19.0
APScheduler==3.10.4
openpyxl==3.1.2
python-dotenv==1.0.0
